chore: Refactor action parsers to use centralized action types and remove unused imports
