+++ packages/stac/lib/src/parsers/actions/stac_delay_action/stac_delay_action_parser.dart
@@ -1,8 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:stac_core/actions/delay/stac_delay_action.dart';
-import 'package:stac/src/utils/action_type.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
 
 class StacDelayActionParser extends StacActionParser<StacDelayAction> {
