+++ packages/stac/lib/src/parsers/actions/stac_form_validate/stac_form_validate_parser.dart
@@ -3,9 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:stac/src/parsers/core/stac_action_parser.dart';
 import 'package:stac/src/parsers/widgets/stac_form/stac_form_scope.dart';
-import 'package:stac/src/utils/action_type.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
-import 'package:stac_core/actions/form_validate/stac_form_validate.dart';
 
 class StacFormValidateParser extends StacActionParser<StacFormValidate> {
   const StacFormValidateParser();
