+++ packages/stac/lib/src/parsers/actions/stac_get_form_value/stac_get_form_value_parser.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:stac/src/parsers/widgets/stac_form/stac_form_scope.dart';
-import 'package:stac/src/utils/action_type.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
-import 'package:stac_core/actions/get_form_value/stac_get_form_value.dart';
 
 class StacGetFormValueParser extends StacActionParser<StacGetFormValue> {
   const StacGetFormValueParser();
