+++ packages/stac/lib/src/parsers/actions/stac_modal_bottom_sheet_action/stac_modal_bottom_sheet_action_parser.dart
@@ -3,10 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:stac/src/framework/framework.dart';
 import 'package:stac/src/parsers/core/stac_widget_parser.dart';
-import 'package:stac_core/actions/modal_bottom_sheet/stac_modal_bottom_sheet_action.dart';
 import 'package:stac/src/parsers/types/type_parser.dart';
-import 'package:stac/src/utils/action_type.dart';
 import 'package:stac/src/utils/color_utils.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
 
 class StacModalBottomSheetActionParser
