+++ packages/stac/lib/src/parsers/actions/stac_multi_action/stac_multi_action_parser.dart
@@ -2,8 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:stac/src/parsers/core/stac_action_parser.dart';
-import 'package:stac_core/actions/multi/stac_multi_action.dart';
-import 'package:stac/src/utils/action_type.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
 
 class StacMultiActionParser extends StacActionParser<StacMultiAction> {
