+++ packages/stac/lib/src/parsers/actions/stac_network_request/stac_network_request_parser.dart
@@ -4,7 +4,6 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:stac/src/framework/framework.dart';
 import 'package:stac/src/services/stac_network_service.dart';
-import 'package:stac/src/utils/action_type.dart';
 import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
 import 'package:stac_logger/stac_logger.dart';
