+++ packages/stac/lib/src/parsers/actions/stac_none_action/stac_none_action_parser.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/widgets.dart';
-import 'package:stac/src/utils/action_type.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
 
 class StacNoneActionParser extends StacActionParser<dynamic> {
