+++ packages/stac/lib/src/parsers/actions/stac_set_value/stac_set_value_action_parser.dart
@@ -1,9 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:stac/src/utils/action_type.dart';
 import 'package:stac/stac.dart';
-import 'package:stac_core/actions/set_value/stac_set_value_action.dart';
+import 'package:stac_core/stac_core.dart';
 
 class StacSetValueActionParser extends StacActionParser<StacSetValueAction> {
   const StacSetValueActionParser();
