+++ packages/stac/lib/src/parsers/actions/stac_snack_bar/stac_snack_bar_parser.dart
@@ -3,12 +3,10 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:stac/src/framework/framework.dart';
 import 'package:stac/src/parsers/core/stac_action_parser.dart';
-import 'package:stac_core/actions/snack_bar/stac_snack_bar.dart';
-import 'package:stac_core/actions/snack_bar/stac_snack_bar_action.dart';
 import 'package:stac/src/parsers/painting/stac_edge_insets_parser.dart';
 import 'package:stac/src/parsers/types/type_parser.dart';
-import 'package:stac/src/utils/action_type.dart';
 import 'package:stac/src/utils/color_utils.dart';
+import 'package:stac_core/stac_core.dart';
 import 'package:stac_framework/stac_framework.dart';
 
 class StacSnackBarParser extends StacActionParser<StacSnackBar> {
