--- packages/stac/lib/src/utils/action_type.dart
@@ -1,13 +1,0 @@
-enum ActionType {
-  navigate,
-  none,
-  networkRequest,
-  showModalBottomSheet,
-  showDialog,
-  getFormValue,
-  validateForm,
-  showSnackBar,
-  setValue,
-  multiAction,
-  delay,
-}
